refactor: split demo into scaffold + math-standards-alignment features
