+++ demos/typescript/README.md
@@ -1,10 +1,14 @@
-# TypeScript SDK Demo — Math Standards Alignment
+# TypeScript SDK Demo
 
 Minimal web app demonstrating the published
 [`@learning-commons/evaluators`](https://www.npmjs.com/package/@learning-commons/evaluators)
-SDK: pick a grade, jurisdiction, and standards, enter a math word problem, and view the
-evaluator's alignment result. React (Vite) frontend + a small Express backend that runs the
-Node-only SDK.
+SDK. The home page lists the available evaluator demos; each is its own page. React (Vite)
+frontend + a small Express backend that runs the Node-only SDK.
+
+Evaluators:
+
+- **Math Standards Alignment** (`/math-standards-alignment`) — pick a grade, jurisdiction, and
+  standards, enter a math word problem, and view the evaluator's alignment result.
 
 ## Setup
 
@@ -24,4 +28,5 @@ cp .env.example .env    # then fill in both keys
 npm run dev
 ```
 
-Open http://localhost:5173 (the Vite dev server proxies `/api` to the backend on `:3001`).
+Open http://localhost:5173 to reach the home page (the Vite dev server proxies `/api` to the
+backend on `:3001`), then pick an evaluator.
