+++ demos/typescript/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Learning Commons — Math Standards Alignment Demo</title>
+    <title>Learning Commons Evaluators — SDK Demo</title>
   </head>
   <body>
     <div id="root"></div>
