+++ demos/typescript/package.json
@@ -15,6 +15,7 @@
     "express": "^5.1.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.1.1",
     "react-select": "^5.10.1"
   },
   "devDependencies": {
