+++ demos/typescript/server/index.ts
@@ -1,75 +1,27 @@
 import 'dotenv/config';
 import express from 'express';
-import { MathStandardsAlignmentEvaluator, Jurisdiction } from '@learning-commons/evaluators';
-import { listStandards } from './kg.js';
-
-const { ANTHROPIC_API_KEY, PLATFORM_API_KEY } = process.env;
-if (!ANTHROPIC_API_KEY || !PLATFORM_API_KEY) {
-  console.error('Missing ANTHROPIC_API_KEY or PLATFORM_API_KEY — copy .env.example to .env and fill in both keys.');
-  process.exit(1);
-}
-
-const evaluator = new MathStandardsAlignmentEvaluator({
-  anthropicApiKey: ANTHROPIC_API_KEY,
-  platformApiKey: PLATFORM_API_KEY,
-});
-
-const GRADES = new Set(['K', '1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']);
-const JURISDICTIONS = new Set<string>(Object.values(Jurisdiction));
-const MAX_QUESTION_LENGTH = 10_000;
+import { registerMathStandardsAlignment } from './math-standards-alignment.js';
 
 const app = express();
 app.use(express.json());
 
-app.get('/api/jurisdictions', (_req, res) => {
-  res.json([...JURISDICTIONS]);
+app.get('/api/health', (_req, res) => {
+  res.json({ status: 'ok' });
 });
 
-app.get('/api/standards', async (req, res) => {
-  const { grade, jurisdiction } = req.query;
-  if (typeof grade !== 'string' || !GRADES.has(grade)) {
-    res.status(400).json({ error: 'grade must be one of K, 1–12' });
-    return;
-  }
-  if (typeof jurisdiction !== 'string' || !JURISDICTIONS.has(jurisdiction)) {
-    res.status(400).json({ error: 'jurisdiction must be a supported Jurisdiction value' });
-    return;
-  }
-  try {
-    res.json(await listStandards(grade, jurisdiction, PLATFORM_API_KEY));
-  } catch (err) {
-    res.status(500).json({ error: err instanceof Error ? err.message : String(err) });
-  }
-});
-
-app.post('/api/evaluate', async (req, res) => {
-  const { question, statementCodes, jurisdiction } = req.body ?? {};
-  if (
-    typeof question !== 'string' ||
-    question.trim().length === 0 ||
-    question.length > MAX_QUESTION_LENGTH ||
-    !Array.isArray(statementCodes) ||
-    statementCodes.length === 0 ||
-    !statementCodes.every((c) => typeof c === 'string' && c.length > 0) ||
-    !JURISDICTIONS.has(jurisdiction)
-  ) {
-    res.status(400).json({
-      error: `question (non-empty, ≤${MAX_QUESTION_LENGTH} chars), statementCodes (non-empty array of strings), and a valid jurisdiction are required`,
-    });
-    return;
-  }
-  try {
-    const results = await evaluator.evaluateItems(
-      [{ question, statementCodes }],
-      jurisdiction as Jurisdiction,
-    );
-    res.json(results);
-  } catch (err) {
-    res.status(500).json({ error: err instanceof Error ? err.message : String(err) });
-  }
-});
+// Each evaluator feature registers its own routes. Add new evaluators here.
+registerMathStandardsAlignment(app);
 
 const PORT = 3001;
-app.listen(PORT, () => {
+const server = app.listen(PORT, () => {
   console.log(`API server listening on http://localhost:${PORT}`);
 });
+
+// Exit promptly on restart/Ctrl-C: close the listener and drop keep-alive
+// sockets so the process doesn't hang until tsx force-kills it after 5s.
+for (const signal of ['SIGINT', 'SIGTERM'] as const) {
+  process.on(signal, () => {
+    server.close(() => process.exit(0));
+    server.closeAllConnections();
+  });
+}
