+++ demos/typescript/src/App.css
@@ -24,6 +24,53 @@ h1 {
   color: #555;
 }
 
+.back {
+  display: inline-block;
+  margin-bottom: 1rem;
+  color: #2563eb;
+  text-decoration: none;
+  font-size: 0.9rem;
+}
+
+.back:hover {
+  text-decoration: underline;
+}
+
+.evaluator-list {
+  list-style: none;
+  padding: 0;
+  margin: 1.5rem 0 0;
+  display: grid;
+  gap: 0.75rem;
+}
+
+.evaluator-list a {
+  display: block;
+  padding: 1rem;
+  background: #fff;
+  border: 1px solid #ddd;
+  border-radius: 6px;
+  text-decoration: none;
+  color: inherit;
+}
+
+.evaluator-list a:hover {
+  border-color: #2563eb;
+}
+
+.evaluator-title {
+  display: block;
+  font-weight: 600;
+  color: #2563eb;
+}
+
+.evaluator-desc {
+  display: block;
+  margin-top: 0.25rem;
+  color: #555;
+  font-size: 0.9rem;
+}
+
 .row {
   display: flex;
   gap: 1rem;
